Cambios Reserva

+++ demo/src/main/java/com/example/demo/repositorios/HabitacionRepositorio.java
@@ -2,12 +2,11 @@ package com.example.demo.repositorios;
 
 import java.util.List;
 import com.example.demo.modelo.Habitacion;
-import com.example.demo.modelo.CategoriaHabitacion;
 import com.example.demo.modelo.EstadoHabitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HabitacionRepositorio extends JpaRepository<Habitacion, Integer> {
+public interface HabitacionRepositorio extends JpaRepository<Habitacion, Long> {
     List<Habitacion> findByEstado(EstadoHabitacion estado);
 }
